+++ belalang_allocator/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "belalang_allocator"
+version = "0.1.0"
+edition = "2021"
+publish = false
+license = "Apache-2.0"
