+++ belalang_allocator/src/lib.rs
@@ -1,5 +1,4 @@
-pub mod stack_allocator;
-
 pub mod block;
 pub mod block_meta;
 pub mod bump_block;
+pub mod stack;
