+++ belalang_vm/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod allocator;
 pub mod builtins;
 pub mod bytecode;
 pub mod error;
